refactor: remove dead code and deduplicate across backend and frontend
